Add CMIP6 Model data process script via upload
This script is written in such a way that it can be used to process C3 and C4 vegetation variables: Area, GPP, carbon content in vegetatin (cVeg) and land (cLand). Here, it is used for the vegetation carbon content. It takes the needed and available variables and calculated the C3 and C4 natural and crop components and their respective totals and save the as netCDF4 and dataframe files.